task4/requirements.txt
functions-framework>=3.0.0
requests>=2.20.0
google-auth>=1.23.0